fix: postApi.ts 복구
